LayoutTests/fast/css/invalid-rule-value.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../js/resources/js-test-pre.js"></script>
</head>
<body>
<div id="test-invalid-rule" style="weight: *; color: rgb(0, 128, 0)"></div>
<script>
description("This test checks that a rule with an invalid value does not halt other style rules from being applied.");

shouldBe("document.getElementById('test-invalid-rule').style.getPropertyValue('color')", "'rgb(0, 128, 0)'");
</script>
<script src="../js/resources/js-test-post.js"></script>
</body>
</html>
